rename "Property" to "Variable" and move into separate file
